chore: move decoration-management as it is not a mixin
