+++ lib/decoration-management.js
@@ -3,7 +3,7 @@
 import { Emitter } from 'atom'
 import { escapeRegExp } from 'underscore-plus'
 import path from 'path'
-import Decoration from '../decoration'
+import Decoration from './decoration'
 
 /**
  * The mixin that provides the decorations API to the minimap editor
