+++ lib/minimap.js
@@ -1,7 +1,7 @@
 'use strict'
 
 import include from './decorators/include'
-import DecorationManagement from './mixins/decoration-management'
+import DecorationManagement from './decoration-management'
 
 import { Emitter, CompositeDisposable } from 'atom'
 import StableAdapter from './adapters/stable-adapter'
